\name{BioSample}
\docType{data}
\alias{BioSample}
\title{Sample from the Fong and Grimmer Wikipedia Biography Data}
\description{This data set gives a small sample of the data used in ``Discovery of Treatments from Text Corpora'' by Christian Fong and Justin Grimmer.  This sample is intended as a toy data set for use in the examples of this package's documentation.  A real data set should include far more observations.}
\usage{BioSample}
\format{A data frame consisting of 51 columns (including an outcome measure and counts for each word in a 50 word vocabulary) and 250 observations.  }
\source{Data collected using the Wikipedia API and an original survey experiment by Fong and Grimmer.}
\references{Fong, Christian and Justin Grimmer. (2016). Discovery of Treatments from Text Corpora. Proceedings of the 54th Annual Meeting of the Association for Computational Linguistics, 1600-1609.}
\keyword{datasets}